fix database url for alembic
