--- alembic/versions/25cf09699af3_.py
@@ -1,67 +1,0 @@
-"""empty message
-
-Revision ID: 25cf09699af3
-Revises: 
-Create Date: 2018-04-30 17:09:05.431717
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = '25cf09699af3'
-down_revision = None
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('locations',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('name', sa.String(), nullable=False),
-    sa.Column('abbreviation', sa.String(), nullable=False),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_table('blocks',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('name', sa.String(), nullable=False),
-    sa.Column('location_id', sa.Integer(), nullable=True),
-    sa.ForeignKeyConstraint(['location_id'], ['locations.id'], ),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_table('floors',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('name', sa.String(), nullable=False),
-    sa.Column('block_id', sa.Integer(), nullable=True),
-    sa.ForeignKeyConstraint(['block_id'], ['blocks.id'], ),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_table('rooms',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('name', sa.String(), nullable=False),
-    sa.Column('room_type', sa.String(), nullable=False),
-    sa.Column('capacity', sa.Integer(), nullable=False),
-    sa.Column('floor_id', sa.Integer(), nullable=True),
-    sa.ForeignKeyConstraint(['floor_id'], ['floors.id'], ),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_table('equipments',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('name', sa.String(), nullable=False),
-    sa.Column('room_id', sa.Integer(), nullable=True),
-    sa.ForeignKeyConstraint(['room_id'], ['rooms.id'], ),
-    sa.PrimaryKeyConstraint('id')
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_table('equipments')
-    op.drop_table('rooms')
-    op.drop_table('floors')
-    op.drop_table('blocks')
-    op.drop_table('locations')
-    # ### end Alembic commands ###
