+++ alembic/versions/286e9c1b8bb9_.py
@@ -1,0 +1,50 @@
+"""empty message
+
+Revision ID: 286e9c1b8bb9
+Revises: 
+Create Date: 2018-05-07 16:22:49.421226
+
+"""
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision = '286e9c1b8bb9'
+down_revision = None
+branch_labels = None
+depends_on = None
+
+
+def upgrade():
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.create_table('equipments',
+    sa.Column('id', sa.Integer(), nullable=False),
+    sa.Column('name', sa.String(), nullable=False),
+    sa.Column('room_id', sa.Integer(), nullable=True),
+    sa.ForeignKeyConstraint(['room_id'], ['rooms.id'], ),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.drop_table('resource')
+    op.drop_table('users')
+    # ### end Alembic commands ###
+
+
+def downgrade():
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.create_table('users',
+    sa.Column('id', sa.INTEGER(), nullable=False),
+    sa.Column('email', sa.VARCHAR(), autoincrement=False, nullable=False),
+    sa.Column('name', sa.VARCHAR(), autoincrement=False, nullable=False),
+    sa.PrimaryKeyConstraint('id', name='users_pkey'),
+    sa.UniqueConstraint('email', name='users_email_key')
+    )
+    op.create_table('resource',
+    sa.Column('id', sa.INTEGER(), nullable=False),
+    sa.Column('name', sa.VARCHAR(), autoincrement=False, nullable=False),
+    sa.Column('room_id', sa.INTEGER(), autoincrement=False, nullable=True),
+    sa.ForeignKeyConstraint(['room_id'], ['rooms.id'], name='resource_room_id_fkey'),
+    sa.PrimaryKeyConstraint('id', name='resource_pkey')
+    )
+    op.drop_table('equipments')
+    # ### end Alembic commands ###
