--- alembic/versions/d70aef59eaa5_added_role.py
@@ -1,40 +1,0 @@
-"""Added role
-
-Revision ID: d70aef59eaa5
-Revises: 5ba451d1fa9d
-Create Date: 2018-04-12 23:54:55.587574
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = 'd70aef59eaa5'
-down_revision = '5ba451d1fa9d'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_table('department')
-    op.add_column('users', sa.Column('role', sa.String(), nullable=True))
-    op.alter_column('users', 'name',
-               existing_type=sa.VARCHAR(length=50),
-               nullable=True)
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.alter_column('users', 'name',
-               existing_type=sa.VARCHAR(length=50),
-               nullable=False)
-    op.drop_column('users', 'role')
-    op.create_table('department',
-    sa.Column('id', sa.INTEGER(), nullable=False),
-    sa.Column('name', sa.VARCHAR(), autoincrement=False, nullable=True),
-    sa.PrimaryKeyConstraint('id', name='department_pkey')
-    )
-    # ### end Alembic commands ###
